Satisfy iOS keyboard review policy
