+++ Packages/iOS/CmuxMobileShellUI/Sources/CmuxMobileShellUI/AgentChatDemoScreenStyle.swift
@@ -1,0 +1,6 @@
+#if DEBUG
+enum AgentChatDemoScreenStyle {
+    case standalone
+    case inlineWorkspace
+}
+#endif
